% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaRC.R
\name{LaRC}
\alias{LaRC}
\title{Landscape Roughness Calculator:}
\usage{
LaRC(data_set, dims = 2, nsize = 16, runs = 100, meth = "lcs")
}
\arguments{
\item{data_set}{tibble containing paired sequences and fitness values}

\item{dims}{number of dimensions requested for landscape}

\item{nsize}{number of data points extracted in each sampling}

\item{runs}{number of times sampling is performed}

\item{meth}{string distance calculation method}
}
\value{
Returns a list object containing the mean 
and standard error of the roughness/slope values obtained
from sampling.
}
\description{
Function for calculating roughness by sampling 
across a fitness landscape based on sequence data. 
Function initially calculates string distances between 
sequences and and condenses the resultant distance
matrix to the number of dimensions requested
a linear fit between the fitness value and the dimensions
is then calculated and the average slope for this fit is c
alculated.The final step is to calculate the sum of the 
residuals to the fit squared divided by the mean slope
to give the roughness/slope.
This calculation is repeated several times with bootstrap 
sampling on the full set and the mean and standard error
of the roughness/slope values is calculated.
}
